cartpole_cl_project/requirements.txt
# Demo1最小依赖
gymnasium>=0.26.0
numpy>=1.21.0
torch>=1.9.0
pyyaml>=6.0
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
pygame>=2.0.0















